- Implement DeleteAsync method in BaseElasticRepository
This commit message signifies the implementation of the DeleteAsync method in the BaseElasticRepository class. The method is responsible for deleting an entity of type T from Elasticsearch. It first checks if the entity is null and throws an exception if so. Then, it calls the CanDelete method to determine if the deletion is allowed, and throws an exception with the provided message if not. The method uses the Elasticsearch client to perform an update operation with a script that deletes the document if the version matches. If the update operation indicates a concurrency exception, an exception with the corresponding localized message is thrown. Finally, the method returns a DeletedResult indicating the outcome of the deletion operation