- Implement InsertAllAsync method in BaseElasticRepository
This commit message signifies the implementation of the InsertAllAsync method in the BaseElasticRepository class. The method is responsible for inserting multiple entities of type T into Elasticsearch in bulk. It first checks if the list of entities is empty and throws an exception if so. Then, it uses the Elasticsearch client to perform a bulk operation, specifying the index name and creating multiple documents from the given entities. The method then processes the response, iterating over each entity and creating an InsertResult based on the success of the corresponding operation. The success is determined by checking if there are any items in the response with errors matching the entity's ID. Finally, the method returns a list of InsertResult objects