- Implement UpdateAsync method in BaseElasticRepository
This commit message indicates the implementation of the UpdateAsync method in the BaseElasticRepository class. The method is responsible for updating an entity of type T in Elasticsearch. It first stores the old version of the entity, then creates a new version of the entity by assigning the current timestamp to UpdateOn and generating a new GUID for Version. The method uses the Elasticsearch client to perform an update operation with a script that checks if the stored version matches the old version and updates the document with the new entity if so. If the update operation results in an error, indicating a concurrency exception, an exception with the corresponding localized message is thrown. Finally, the method returns an UpdatedResult based on the update result, generated by invoking the ToUpdateResult extension method on the entity