- Implement IncrementAsync method in BaseElasticRepository
This commit message indicates the implementation of the IncrementAsync method in the BaseElasticRepository class. The method is responsible for incrementing a specific field of an entity in Elasticsearch. It uses the Elasticsearch client to perform an update operation with a script that checks if the stored version matches the provided version. If the versions match, the script increments the specified field by the given amount. If the versions do not match, an exception with the message "Invalid version" is thrown. The method then checks if the update operation resulted in an error, indicating a concurrency exception, and throws an exception with the corresponding localized message if so. Finally, the method returns an IncrementResult indicating the success of the operation, based on the validity of the response