- Implement FirstOrDefaultAsync method in BaseElasticRepository
This commit message signifies the implementation of the FirstOrDefaultAsync method in the BaseElasticRepository class. The method is responsible for retrieving the first document of type T from Elasticsearch. It uses the Elasticsearch client to perform a search operation with a match-all query, limiting the size of the response to 1 and disabling tracking of total hits. The method then returns the first document from the response, using the FirstOrDefault method